% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeMarkov.R
\name{quality}
\alias{quality}
\title{quality}
\usage{
quality(file, pos_training = NULL, neg_training = NULL,
  trans_pos = NULL, trans_neg = NULL, l_word_pos = 1,
  l_word_neg = 1, n_train = 1, n_seq = 1, quiet = FALSE)
}
\arguments{
\item{file}{a file (fasta) to read and use to test model}

\item{pos_training}{a file (fasta) to read and train the positive model}

\item{neg_training}{a file (fasta) to read and train the negative model}

\item{trans_pos}{a transition matrix if it was already computed. Therefore, no need to train models. Warning, it must be in log}

\item{trans_neg}{a transition matrix if it was already computed. Therefore, no need to train models. Warning, it must be in log}

\item{l_word_pos}{length of words for the model. Equal to the "order of the model + 1"}

\item{l_word_neg}{length of words for the model. Equal to the "order of the model + 1"}

\item{n_train}{number of sequences to train with}

\item{n_seq}{number of sequences to analyse}

\item{quiet}{if some informations are print or not (boolean)}
}
\description{
Compute the quality of the model and return a TRUE Positive or FALSE Negative information. 
It test the positive and negative model and assign every sequence of the fasta file to one model or the other. 
Therefore, the file must countain sequences which are know to be from one model.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
