% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeMarkov.R
\name{smoothing}
\alias{smoothing}
\title{smoothing}
\usage{
smoothing(seq, l_word_pos = 5, l_word_neg = 4, smooth_win = 10,
  reject_win = 1)
}
\arguments{
\item{seq}{export table of viterbi function}

\item{l_word_pos}{a value for lengths of words for the model. Equal to the "order of the model + 1"}

\item{l_word_neg}{a value fof lengths of words for the model. Equal to the "order of the model + 1"}

\item{smooth_win}{minimal region length for a region to be kept under model + or -}

\item{reject_win}{minimal region length for a region ambiguous to be kept under model ambiguous}
}
\description{
Use the export table of viterbi function and apply two algorithms to smooth the data by two different variables
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
