% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeMarkov.R
\name{threshold}
\alias{threshold}
\title{threshold}
\usage{
threshold(pos_test, neg_test, pos_training, neg_training,
  pos_seq = c(1:2), neg_seq = c(1:2), n_train = 1, n_seq = 1)
}
\arguments{
\item{pos_test}{a file (fasta) to read and use to test positive model}

\item{neg_test}{a file (fasta) to read and use to test negative model}

\item{pos_training}{a file (fasta) to read and train the positive model}

\item{neg_training}{a file (fasta) to read and train the negative model}

\item{pos_seq}{a vector for lengths of words for the model. Equal to the "order of the model + 1"}

\item{neg_seq}{a vector fof lengths of words for the model. Equal to the "order of the model + 1"}

\item{n_train}{number of sequences to train with}

\item{n_seq}{number of sequences to analyse}
}
\description{
Compute the quality for multiple set of values in vectors. 
Compute TRUE Positive or FALSE Negative information and FALSE Positive and TRUE Negative. 
All of this allow to compute sensitivoty and specificity of every model.
 
Compute the quality of the model and return a TRUE Positive or FALSE Negative information. 
It test the positive and negative model and assign every sequence of the fasta file to one model or the other. 
Therefore, the file must countain sequences which are know to be from one model.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
