% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeMarkov.R
\name{transition}
\alias{transition}
\title{transition}
\usage{
transition(file, l_word = 1, n_seq = 1, log = TRUE, type = "")
}
\arguments{
\item{file}{a file (fasta) to read and use as training for model}

\item{l_word}{length of words for the model. Equal to the "order of the model + 1"}

\item{n_seq}{number of sequences to train the model with}

\item{log}{boolean information if the transition matrix of the model is in log or not.}

\item{type}{a indicator ("+" or "-") for printing pretty stuff during computations}
}
\description{
Compute the transition model for one Markov model on DNA.
}
\author{
Jaunatre Maxime <maxime.jaunatre@etu.univ-grenoble-alpes.fr>
}
